pyproject.toml
[project]
name = "arena-evals"
version = "0.1.0"
description = "Worked solutions to ARENA Chapter 3"
readme = "README.md"
authors = [
    { name = "Matt Clarke", email = "mattjclarke94@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.45.0",
    "httpx>=0.28.1",
    "marimo[sql]>=0.10.17",
    "openai>=1.60.1",
    "polars>=1.21.0",
    "python-dotenv>=1.0.1",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
